chore(browseros): move release and ota toolsets out of steps
