+++ packages/browseros/bos_build/release/appcast.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3
 """Appcast module - Generate Sparkle appcast XML snippets"""
 
-from ...core.context import Context
-from ...core.step import Step, ValidationError
-from ...lib.utils import log_info, log_warning
-from ..storage import BOTO3_AVAILABLE
+from ..core.context import Context
+from ..core.step import Step, ValidationError
+from ..lib.utils import log_info, log_warning
+from ..lib.r2 import BOTO3_AVAILABLE
 from .common import fetch_all_release_metadata, generate_appcast_item
