+++ packages/browseros/bos_build/release/github.py
@@ -6,10 +6,10 @@ import tempfile
 from pathlib import Path
 from typing import Dict, List, Optional, Tuple
 
-from ...core.context import Context
-from ...core.step import Step, ValidationError
-from ...lib.utils import log_info, log_error, log_success, log_warning
-from ..storage import BOTO3_AVAILABLE
+from ..core.context import Context
+from ..core.step import Step, ValidationError
+from ..lib.utils import log_info, log_error, log_success, log_warning
+from ..lib.r2 import BOTO3_AVAILABLE
 from .common import (
     PLATFORMS,
     PLATFORM_DISPLAY_NAMES,
